requirements.txt
humanize
fastapi
uvicorn[standard]
pillow
aiofiles
git+https://github.com/reupen/aiogithub.git#egg=aiogithub